account: incorporate `formBase` functionality
